Major changes to fragment_home.xml file
Added scroll view. Added two linear layout for upcoming bill and overdue bills respectively. The two linear layout will act as the container of the subscriptions seen in the home page. This is also were the card view be shown (contents of the subscription under upcoming bill or overdue bills)